% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getATE_posterior_NDPM.R
\name{getATE_posterior_NDPM}
\alias{getATE_posterior_NDPM}
\title{getATE_posterior_NDPM}
\usage{
getATE_posterior_NDPM(fittedModel, X, niter, burnin = 0, thin = 1)
}
\arguments{
\item{fittedModel}{: model fits of the class 'BayesIV'}

\item{X}{: a numeric matrix of covariates (excluding the column with intercept)}

\item{niter}{: number indicating the size of the MCMC sample}

\item{thin}{: number to thin the MCMC chain with}

\item{burnin:}{number indicating burnin of the MCMC chain}
}
\value{
a list with posterior draws of Y1hat, Y0hat and ATE
}
\description{
getATE_posterior_NDPM
}
\author{
Sam Adhikari
}
