% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTT_posterior.R
\name{getTT_posterior}
\alias{getTT_posterior}
\title{getTT_posterior}
\usage{
getTT_posterior(fittedModel, X, niter, burnin = 0, thin = 1)
}
\arguments{
\item{fittedModel}{: model fits of the class 'BayesIV'}

\item{X}{: a numeric matrix of covariates (excluding the column with intercept)}

\item{niter}{: number indicating the size of the MCMC sample}

\item{thin}{: number to thin the MCMC chain with}

\item{burnin:}{number indicating burnin of the MCMC chain}
}
\value{
posterior chain of ATT
}
\description{
getTT_posterior
}
