% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTT_posterior_NDPM.R
\name{getTT_posterior_NDPM}
\alias{getTT_posterior_NDPM}
\title{getTT_posterior_NDPM}
\usage{
getTT_posterior_NDPM(fittedModel, XOc, XTr, Z, niter, burnin = 0, thin = 1)
}
\arguments{
\item{fittedModel}{: model fits of the class 'BayesIV'}

\item{XTr}{: a numeric matrix of covariates}

\item{niter}{: number indicating the size of the MCMC sample}

\item{thin}{: number to thin the MCMC chain with}

\item{X0c}{: a numeric matrix of covariates (excluding the column with intercept)}

\item{Z:}{A vector of instrumental variable}

\item{burnin:}{number indicating burnin of the MCMC chain}
}
\value{
posterior chain of ATT
}
\description{
getTT_posterior_NDPM
}
\author{
Sam Adhikari
}
