% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMTE_posterior.R
\name{getweights_MTE}
\alias{getweights_MTE}
\title{getMTE_posterior}
\usage{
getweights_MTE(fittedModel, u, niter, burnin = 0, thin = 1)
}
\arguments{
\item{fittedModel}{: model fits of the class 'BayesIV'}

\item{niter}{: number indicating the size of the MCMC sample}

\item{thin}{: number to thin the MCMC chain with}

\item{u:}{margin}

\item{burnin:}{number indicating burnin of the MCMC chain}
}
\value{
a list with posterior draws of MTE
}
\description{
getMTE_posterior
}
\author{
Sam Adhikari
}
