% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gibbs_Reg.R
\name{Gibbs_Reg}
\alias{Gibbs_Reg}
\title{Gibbs_Reg: Gibbs sampler function to sample slope coefficients in
linear regression with normal noise}
\usage{
Gibbs_Reg(Y, X, sigmasq_prior, sigmasq_prior_alpha, sigmasq, PP)
}
\arguments{
\item{Y}{: dependent variable}

\item{X}{: covariates}

\item{sigmasq_prior}{: variance of normal prior}

\item{sigmasq_prior_alpha}{: variance of the normal prior for \code{alpha}}

\item{sigmasq}{: variance of \code{Y}}

\item{PP}{: number of covariates}
}
\value{
A draw from the posterior distribution
}
\description{
Gibbs_Reg: Gibbs sampler function to sample slope coefficients in
linear regression with normal noise
}
