% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_NormalLatentFactor.R
\name{mcmcRun_Normal}
\alias{mcmcRun_Normal}
\title{mcmcRun_Normal: MCMC sampler for IV Analysis with Normal prior on the latent factors}
\usage{
mcmcRun_Normal(Yobs, Tr, X, Z, niter)
}
\arguments{
\item{Tr}{: Binary numeric vector of treatment indicator.}

\item{X}{: numeric matrix of covariates of dimension n-By-p}

\item{Z}{: numeric vector of instrumental variable of length n}

\item{niter}{: number of MCMC sampler to be run}

\item{Yobs:}{Input numeric vector of observed outcome of length n, can be binary or continuous.}
}
\value{
runModel : list of posterior samples for each parameter
}
\description{
mcmcRun_Normal: MCMC sampler for IV Analysis with Normal prior on the latent factors
}
\examples{
\dontrun{obj = mcmcRun_Normal(Yobs,Tr,X,Z,niter)}
}
\author{
: Sam Adhikari
}
