% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_IV_Normal_DPM.R
\name{mcmcRun_Normal_DPM}
\alias{mcmcRun_Normal_DPM}
\title{mcmcRun_Normal_DPM: MCMC sampler for IV Analysis with Dirichlet process mixture prior on the errors and Normal prior on the latent factors}
\usage{
mcmcRun_Normal_DPM(Yobs, Tr, X, Z, niter, priors = NULL, initialVals = NULL)
}
\arguments{
\item{Yobs}{: Input numeric vector of observed outcome of length n, can be binary or continuous.}

\item{Tr}{: Binary numeric vector of treatment indicator.}

\item{X}{: numeric matrix of covariates of dimension n-By-p}

\item{Z}{: numeric vector of instrumental variable of length n}

\item{niter}{: number of MCMC sampler to be run}

\item{priors:}{list to specify the parameters for the prior distribution. If NULL, default specification is used.}

\item{initialVals:}{list of initial values for each parameter. If NULL, default setting for initialization is used.}
}
\value{
runModel : list of posterior samples for each parameter
}
\description{
mcmcRun_Normal_DPM: MCMC sampler for IV Analysis with Dirichlet process mixture prior on the errors and Normal prior on the latent factors
}
\examples{
{ 
obj = mcmcRun_Normal_DPM(Yobs=Yobs,Tr=D,X=X[,-1],Z=Z,niter=10,priors=NULL,initialVals=NULL)
}
}
\author{
Sam Adhikari
}
