% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_IV_DirichletLatentFactor.R
\name{mcmcRunDP}
\alias{mcmcRunDP}
\title{mcmcRunDP: MCMC sampler for IV Analysis with Dirichlet process mixture prior on the latent factors}
\usage{
mcmcRunDP(Yobs, Tr, X, Z, niter)
}
\arguments{
\item{Tr}{: Binary numeric vector of treatment indicator.}

\item{X}{: numeric matrix of covariates of dimension n-By-p}

\item{Z}{: numeric vector of instrumental variable of length n}

\item{niter}{: number of MCMC sampler to be run}

\item{Yobs:}{Input numeric vector of observed outcome of length n, can be binary or continuous.}
}
\value{
runModel : list of posterior samples for each parameter, acceptance rate and tuning parameter
}
\description{
mcmcRunDP: MCMC sampler for IV Analysis with Dirichlet process mixture prior on the latent factors
}
\examples{
\dontrun{obj = mcmcRunDP(Yobs,Tr,X,Z,niter)}
}
\author{
Sam Adhikari
}
