% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotATEs.R
\name{plotATEs}
\alias{plotATEs}
\title{plotATEs: Function to plot posterior chain of ATEs}
\usage{
plotATEs(xx, pdfname = NULL, int.quantiles, effect_true, effect_obs,
  labels.axis = NULL)
}
\description{
plotATEs: Function to plot posterior chain of ATEs
}
